backend/fish_store/requirements.txt
Django==4.2.6
djangorestframework==3.14.0
django-cors-headers==4.3.0
psycopg2-binary==2.9.9
django-extensions==3.2.3